Make successful jar file
