+++ data/data.txt
@@ -1,0 +1,3 @@
+T | 0 | d1
+T | 0 | d2
+T | 0 | d3
